Add static way of getting data
This is the only way I could really find so far. Trying to use an API usually gives me CORS errors (CORS header 'Access-Control-Allow-Origin' missing).

To fix this, I made a mildly annoying way to setup it up, and it's static, but once it works, it's great.

This will allow users to use the website without having to clone the project itself.