Edit design bug
This website has 3 main designs. When viewed from a phone, it should show no background, just a gray screen with the necessary information.
When the phone / device is big enough, it will show the background with the main design applied, but a couple of tweaks, like with the logo, that should appear in the middle.

Until now, that didn't happen, and that was an oversight made by me.
While experimenting on the experiment branch, I figured what the problem was, so I applied it there.

I do think it should be inmediately applied, hence why I pushed this change directly to master.

This should fix it, for the most part.